LayoutTests/imported/w3c/web-platform-tests/css/css-tables/bounding-box-computation-2-expected.txt

Bounding box computation

Checks that all table-internal boxes are positioned correctly

The all table-internal elements of this table shoud be 100px and 100px wide tall:
Table size is 120px. There are two 10px borders, one at each side.


PASS Control test: Table width is 120px
PASS Control test: Table height is 120px
PASS Table-cell is 100px wide
PASS Table-cell is 100px tall
PASS Table-row is 100px wide
PASS Table-row is 100px tall
PASS Table-row-group is 100px wide
PASS Table-row-group is 100px tall
PASS Table-column is 100px wide
PASS Table-column is 100px tall
PASS Table-column-group is 100px wide
PASS Table-column-group is 100px tall
